monitoramento2/src/requirements.txt
psycopg2-binary
ping3
requests